[fix] Fixed a bug that created very high latency when not using a separate ImGui viewport.
- It turns out VkAcquireNextImageKHR(...) does not perform as expected in FIFO modes
(The function does not block when no image is directly available (up to a certain extent), but leaves it to the semaphore/fence to signal the availability.
So even with a FIFO with 2 images, we can acquire many (maybe 30+) images in advance before the function blocks, leading to massive latency.
To fix it, we need to manually wait for images fences in FIFO modes.)
- Added a lot of logging to measure timings.
- Added an AsynchTask global mutex to not lock g_common_mutex (renamed from g_mutex) when a task is waiting user input.